Users loaded correctly to table

+++ Objective-C UI Project/Source/FifthTask/Models/FaceBookUsers/ANSContexts/ANSFBLoginContext.m
@@ -34,7 +34,7 @@
     }
     
     NSString *ID = (NSString *)[result objectForKeyedSubscript:@"id"];
-    user.ID = ID.intValue;
+    user.ID = ID.doubleValue;
     user.state = ANSUserDidLoadID;
 }
